+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - actor sheets now only scroll their tabbed content instead of the entire sheet.
   This means that the stats header is always visible.
 - further condensed the actor sheet item table
+- only effect items' rule elements are directly applied to the actor for now.
+  Weapon and apparel rule elements will soon only apply when equipped
 
 ### Fixed